Add quest

+++ src/NosCore.Parser/Parsers/NpcMonsterParser.cs
@@ -19,10 +19,7 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Text;
-using Autofac.Core;
 using NosCore.Core;
 using NosCore.Core.I18N;
 using NosCore.Data.Enumerations.I18N;
@@ -450,9 +447,9 @@ namespace NosCore.Parser.Parsers
 
         }
 
-        private short ImportHeroXp(Dictionary<string, string[][]> chunk)
+        private int ImportHeroXp(Dictionary<string, string[][]> chunk)
         {
-            return (short)(Convert.ToInt16(chunk["VNUM"][0][2]) switch
+            return (Convert.ToInt32(chunk["VNUM"][0][2]) switch
             {
                 2510 => 881,
                 2501 => 881,
